+++ payment-request.debug.js
@@ -1145,7 +1145,7 @@ module.exports = {
 var BraintreeError = _dereq_('./braintree-error');
 var Promise = _dereq_('./promise');
 var sharedErrors = _dereq_('./errors');
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 
 function basicComponentVerification(options) {
   var client, authorization, name;
@@ -1447,7 +1447,7 @@ module.exports = BraintreeBus;
 },{"../braintree-error":20,"./check-origin":21,"./events":22,"framebus":13}],24:[function(_dereq_,module,exports){
 'use strict';
 
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 var PLATFORM = 'web';
 
 var CLIENT_API_URLS = {
@@ -1574,7 +1574,7 @@ var Promise = _dereq_('./promise');
 var assets = _dereq_('./assets');
 var sharedErrors = _dereq_('./errors');
 
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 
 function createDeferredClient(options) {
   var promise = Promise.resolve();
@@ -1687,7 +1687,7 @@ module.exports = {
 },{"./braintree-error":20}],31:[function(_dereq_,module,exports){
 'use strict';
 
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 var assign = _dereq_('./assign').assign;
 
 function generateTokenizationParameters(configuration, overrides) {
@@ -1946,7 +1946,7 @@ var methods = _dereq_('../../lib/methods');
 var Promise = _dereq_('../../lib/promise');
 var EventEmitter = _dereq_('@braintree/event-emitter');
 var BraintreeError = _dereq_('../../lib/braintree-error');
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 var constants = _dereq_('../shared/constants');
 var events = constants.events;
 var errors = constants.errors;
@@ -2634,7 +2634,7 @@ var basicComponentVerification = _dereq_('../lib/basic-component-verification');
 var createDeferredClient = _dereq_('../lib/create-deferred-client');
 var createAssetsUrl = _dereq_('../lib/create-assets-url');
 var wrapPromise = _dereq_('@braintree/wrap-promise');
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 
 /**
  * @static