+++ preferred-payment-methods.debug.js
@@ -676,7 +676,7 @@ module.exports = {
 var BraintreeError = _dereq_('./braintree-error');
 var Promise = _dereq_('./promise');
 var sharedErrors = _dereq_('./errors');
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 
 function basicComponentVerification(options) {
   var client, authorization, name;
@@ -806,7 +806,7 @@ module.exports = BraintreeError;
 },{"./enumerate":18}],14:[function(_dereq_,module,exports){
 'use strict';
 
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 var PLATFORM = 'web';
 
 var CLIENT_API_URLS = {
@@ -915,7 +915,7 @@ var Promise = _dereq_('./promise');
 var assets = _dereq_('./assets');
 var sharedErrors = _dereq_('./errors');
 
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 
 function createDeferredClient(options) {
   var promise = Promise.resolve();
@@ -1093,7 +1093,7 @@ module.exports = {
 var wrapPromise = _dereq_('@braintree/wrap-promise');
 var basicComponentVerification = _dereq_('../lib/basic-component-verification');
 var PreferredPaymentMethods = _dereq_('./preferred-payment-methods');
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 
 /**
  * @static