+++ venmo.debug.js
@@ -783,7 +783,7 @@ module.exports = {
 var BraintreeError = _dereq_('./braintree-error');
 var Promise = _dereq_('./promise');
 var sharedErrors = _dereq_('./errors');
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 
 function basicComponentVerification(options) {
   var client, authorization, name;
@@ -913,7 +913,7 @@ module.exports = BraintreeError;
 },{"./enumerate":28}],23:[function(_dereq_,module,exports){
 'use strict';
 
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 var PLATFORM = 'web';
 
 var CLIENT_API_URLS = {
@@ -1040,7 +1040,7 @@ var Promise = _dereq_('./promise');
 var assets = _dereq_('./assets');
 var sharedErrors = _dereq_('./errors');
 
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 
 function createDeferredClient(options) {
   var promise = Promise.resolve();
@@ -1328,7 +1328,7 @@ var BraintreeError = _dereq_('../lib/braintree-error');
 var Venmo = _dereq_('./venmo');
 var Promise = _dereq_('../lib/promise');
 var supportsVenmo = _dereq_('./shared/supports-venmo');
-var VERSION = "3.61.0";
+var VERSION = "3.62.0";
 
 /**
  * @static
@@ -1337,6 +1337,7 @@ var VERSION = "3.61.0";
  * @param {Client} [options.client] A {@link Client} instance.
  * @param {string} [options.authorization] A tokenizationKey or clientToken. Can be used in place of `options.client`.
  * @param {boolean} [options.allowNewBrowserTab=true] This should be set to false if your payment flow requires returning to the same tab, e.g. single page applications. Doing so causes {@link Venmo#isBrowserSupported|isBrowserSupported} to return true only for mobile web browsers that support returning from the Venmo app to the same tab.
+ * @param {boolean} [options.ignoreHistoryChanges=false] When the Venmo app returns to the website, it will modify the hash of the url to include data about the tokenization. By default, the SDK will put the state of the hash back to where it was before the change was made. Pass `true` to handle the hash change instead of the SDK.
  * @param {string} [options.profileId] The Venmo profile ID to be used during payment authorization. Customers will see the business name and logo associated with this Venmo profile, and it will show up in the Venmo app as a "Connected Merchant". Venmo profile IDs can be found in the Braintree Control Panel. Omitting this value will use the default Venmo profile.
  * @param {string} [options.deepLinkReturnUrl] An override for the URL that the Venmo iOS app opens to return from an app switch.
  * @param {callback} [callback] The second argument, `data`, is the {@link Venmo} instance. If no callback is provided, `create` returns a promise that resolves with the {@link Venmo} instance.
@@ -1583,6 +1584,7 @@ function Venmo(options) {
   this._allowNewBrowserTab = options.allowNewBrowserTab !== false;
   this._profileId = options.profileId;
   this._deepLinkReturnUrl = options.deepLinkReturnUrl;
+  this._ignoreHistoryChanges = options.ignoreHistoryChanges;
 }
 
 Venmo.prototype.getUrl = function () {
@@ -1644,7 +1646,13 @@ Venmo.prototype.isBrowserSupported = function () {
  * @returns {boolean} True if the results of Venmo payment authorization are available and ready to process.
  */
 Venmo.prototype.hasTokenizationResult = function () {
-  var params = getFragmentParameters();
+  return this._hasTokenizationResult();
+};
+
+// a private version that lets us pass in a custom hash
+// when listening on a hashchange event
+Venmo.prototype._hasTokenizationResult = function (hash) {
+  var params = getFragmentParameters(hash);
 
   return typeof (params.venmoSuccess || params.venmoError || params.venmoCancel) !== 'undefined';
 };
@@ -1689,6 +1697,7 @@ Venmo.prototype.hasTokenizationResult = function () {
  */
 Venmo.prototype.tokenize = function (options) {
   var self = this;
+  var resultProcessingInProgress, visibilityChangeListenerTimeout;
 
   options = options || {};
 
@@ -1703,38 +1712,70 @@ Venmo.prototype.tokenize = function (options) {
   this._tokenizationInProgress = true;
   this._tokenizePromise = new ExtendedPromise();
 
-  // Subscribe to document visibility change events to detect when app switch
-  // has returned.
   this._previousHash = global.location.hash;
-  this._visibilityChangeListener = function () {
+
+  function completeFlow(hash) {
     var error;
+
+    self._processResults(hash).catch(function (err) {
+      error = err;
+    }).then(function (res) {
+      if (!self._ignoreHistoryChanges && global.location.hash !== self._previousHash) {
+        global.location.hash = self._previousHash;
+      }
+      self._removeVisibilityEventListener();
+
+      if (error) {
+        self._tokenizePromise.reject(error);
+      } else {
+        self._tokenizePromise.resolve(res);
+      }
+      delete self._tokenizePromise;
+    });
+  }
+
+  // The Venmo SDK app switches back with the results of the
+  // tokenization encoded in the hash
+  this._onHashChangeListener = function (e) {
+    var hash = e.newURL.split('#')[1];
+
+    if (!self._hasTokenizationResult(hash)) {
+      return;
+    }
+
+    resultProcessingInProgress = true;
+    clearTimeout(visibilityChangeListenerTimeout);
+    self._tokenizationInProgress = false;
+    completeFlow(hash);
+  };
+
+  window.addEventListener('hashchange', this._onHashChangeListener, false);
+
+  // Subscribe to document visibility change events to detect when app switch
+  // has returned. Acts as a fallback for the hashchange listener and catches
+  // the cancel case via manual app switch back
+  this._visibilityChangeListener = function () {
+    var delay = options.processResultsDelay || constants.DEFAULT_PROCESS_RESULTS_DELAY;
 
     if (!global.document.hidden) {
       self._tokenizationInProgress = false;
 
-      setTimeout(function () {
-        self._processResults().catch(function (err) {
-          error = err;
-        }).then(function (res) {
-          global.location.hash = self._previousHash;
-          self._removeVisibilityEventListener();
-          delete self._visibilityChangeListener;
-
-          if (error) {
-            self._tokenizePromise.reject(error);
-          } else {
-            self._tokenizePromise.resolve(res);
-          }
-          delete self._tokenizePromise;
-        });
-      }, options.processResultsDelay || constants.DEFAULT_PROCESS_RESULTS_DELAY);
+      if (!resultProcessingInProgress) {
+        visibilityChangeListenerTimeout = setTimeout(completeFlow, delay);
+      }
     }
   };
 
   return this.getUrl().then(function (url) {
-    // Deep link URLs do not launch iOS apps from a webview when using window.open or PopupBridge.open.
     if (self._deepLinkReturnUrl) {
-      global.location = url;
+      if (isIosWebview()) {
+        // Deep link URLs do not launch iOS apps from a webview when using window.open or PopupBridge.open.
+        global.location.href = url;
+      } else if (global.popupBridge && typeof global.popupBridge.open === 'function') {
+        global.popupBridge.open(url);
+      } else {
+        global.open(url);
+      }
     } else {
       global.open(url);
     }
@@ -1768,12 +1809,16 @@ Venmo.prototype.teardown = function () {
 };
 
 Venmo.prototype._removeVisibilityEventListener = function () {
+  global.removeEventListener('hashchange', this._onHashChangeListener);
   global.document.removeEventListener(documentVisibilityChangeEventName(), this._visibilityChangeListener);
+
+  delete this._visibilityChangeListener;
+  delete this._onHashChangeListener;
 };
 
-Venmo.prototype._processResults = function () {
+Venmo.prototype._processResults = function (hash) {
   var self = this;
-  var params = getFragmentParameters();
+  var params = getFragmentParameters(hash);
 
   return new Promise(function (resolve, reject) {
     if (params.venmoSuccess) {
@@ -1801,12 +1846,22 @@ Venmo.prototype._processResults = function () {
       reject(new BraintreeError(errors.VENMO_CANCELED));
     }
 
-    clearFragmentParameters();
+    self._clearFragmentParameters();
   });
 };
 
-function getFragmentParameters() {
-  var keyValuesArray = global.location.hash.substring(1).split('&');
+Venmo.prototype._clearFragmentParameters = function () {
+  if (this._ignoreHistoryChanges) {
+    return;
+  }
+
+  if (typeof global.history.replaceState === 'function' && global.location.hash) {
+    history.pushState({}, '', global.location.href.slice(0, global.location.href.indexOf('#')));
+  }
+};
+
+function getFragmentParameters(hash) {
+  var keyValuesArray = (hash || global.location.hash.substring(1)).split('&');
 
   return keyValuesArray.reduce(function (toReturn, keyValue) {
     var parts = keyValue.split('=');
@@ -1821,12 +1876,6 @@ function getFragmentParameters() {
 
     return toReturn;
   }, {});
-}
-
-function clearFragmentParameters() {
-  if (typeof global.history.replaceState === 'function' && global.location.hash) {
-    history.pushState({}, '', global.location.href.slice(0, global.location.href.indexOf('#')));
-  }
 }
 
 function formatTokenizePayload(fragmentParams) {
@@ -1852,6 +1901,14 @@ function documentVisibilityChangeEventName() {
   }
 
   return visibilityChange;
+}
+
+function isIosWebview() {
+  // we know it's a webview because this flow only gets
+  // used when checking the deep link flow
+  // test the platform here to get around custom useragents
+  return global.navigator.platform &&
+    /iPhone|iPad|iPod/.test(global.navigator.platform);
 }
 
 module.exports = wrapPromise.wrapPrototype(Venmo);